refactor: move toggle state out of CredentialContext
